Added LinkBuilder class/implicits to convert Urls to hyperlinks.
The `bigtop.core.LinkBuilder` class is constructed with a `bigtop.core.Url` and contains a couple of `link()` methods to generate hyperlinks (instances of `scala.xml.Elem`).

There is an implicit in the `bigtop.core` package to promote a `Url` to a `LinkBuilder`. This allows you to do things like this:

    Url("http://example.com").path(List("foo")).link("Click here")  // ==> <a href="http://example.com/foo">Click here</a>

and this if you're using Bigtop Routes (example taken from the bigtop demo site):

    Calculator.add.url(1, 2).link("Calculate!", "class" -> "calc-link") // ==> <a href="/add/1/to/2" title="calc-link">Calculate!</a>