Added a mechanism to retrieve request/response data in Routes.
`bigtop.routes.core.Site` now contains a `request` method that returns a framework-specific request object (e.g. `Req` in Lift, `HttpServletRequest` in a plain servlet). The method result is only defined within the dynamic context of a call to `Site.apply()`.

The plain servlet version of `Site` is now complete. Because Java servlets accumulate response data via mutation, the plain site and routes are implemented as functions of type `(HttpServletRequest) => Unit`, rather than `(HttpServletRequest) => HttpServletResponse` as might seem more natural.

The plain `Site` object contains a `response` method in addition to `request`, which returns the `HttpServletResponse` currently being accumulated. The top-level entry-point to `Site` is a two-argument `apply(HttpServletRequest, HttpServletResponse)` method that sets up the request and response object and invokes the rest of the dispatching process.